Move mouse implementations to own cpp
